tests/backend/requirements-test.txt
# Python测试依赖

# 核心测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-html>=3.2.0
pytest-metadata>=3.0.0

# Mock和测试工具
responses>=0.24.0
requests-mock>=1.11.0
factory-boy>=3.3.0
faker>=20.0.0
freezegun>=1.2.0

# 代码质量
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# 类型检查
types-PyYAML>=6.0.0
types-requests>=2.31.0

# 报告生成
allure-pytest>=2.13.0
